+++ allennlp/training/callbacks/log_writer.py
@@ -1,40 +1,30 @@
 from collections import deque, defaultdict
-from typing import Any, Dict, List, Optional, Set, Union, Deque
 import logging
+from typing import List, Dict, Any, Optional, TYPE_CHECKING, Union, Deque, Set
 
 import torch
 
-from allennlp.common import Registrable
 from allennlp.data import TensorDict
-from allennlp.nn import util as nn_util
-from allennlp.training.optimizers import Optimizer
-from allennlp.training import util as training_util
-from allennlp.models.model import Model
+from allennlp.nn.util import tiny_value_of_dtype
+from allennlp.training.callbacks.callback import TrainerCallback
+from allennlp.training.util import get_train_and_validation_metrics, get_batch_size
+
+if TYPE_CHECKING:
+    from allennlp.training.trainer import GradientDescentTrainer
 
 
 logger = logging.getLogger(__name__)
 
 
-class LogWriter(Registrable):
+class LogWriterCallback(TrainerCallback):
     """
-    Class that handles trainer logging.
+    An abstract baseclass for callbacks that Log training statistics and metrics.
+    Examples of concrete implementations are the `TensorBoardCallback` and `WandBCallback`.
 
     # Parameters
 
     serialization_dir : `str`
         The training serialization directory.
-
-        In a typical AllenNLP configuration file, this parameter does not get an entry in the
-        file, it gets passed in separately.
-
-    model : `Model`
-        The model being trained.
-
-        In a typical AllenNLP configuration file, this parameter does not get an entry in the
-        file, it gets passed in separately.
-
-    optimizer : `Optimizer`
-        The optimizer used to train the model.
 
         In a typical AllenNLP configuration file, this parameter does not get an entry in the
         file, it gets passed in separately.
@@ -73,24 +63,19 @@ class LogWriter(Registrable):
 
     batch_loss_moving_average_count : `int`, optional (default = `100`)
         The length of the moving average for batch loss.
-
     """
 
     def __init__(
         self,
         serialization_dir: str,
-        model: Model,
-        optimizer: Optimizer,
         summary_interval: int = 100,
         distribution_interval: Optional[int] = None,
         batch_size_interval: Optional[int] = None,
         should_log_parameter_statistics: bool = True,
         should_log_learning_rate: bool = False,
         batch_loss_moving_average_count: int = 100,
-    ):
-        self._serialization_dir = serialization_dir
-        self._model = model
-        self._optimizer = optimizer
+    ) -> None:
+        super().__init__(serialization_dir)
         self._summary_interval = summary_interval
         self._distribution_interval = distribution_interval
         self._batch_size_interval = batch_size_interval
@@ -99,12 +84,10 @@ class LogWriter(Registrable):
         self._cumulative_batch_group_size = 0
         self._distribution_parameters: Optional[Set[str]] = None
         self._module_hook_handles: List[torch.utils.hooks.RemovableHandle] = []
-        self._batch_num_total: int = 0
         self._batch_loss_moving_average_count = batch_loss_moving_average_count
         self._batch_loss_moving_sum: Dict[str, float] = defaultdict(float)
         self._batch_loss_moving_items: Dict[str, Deque[float]] = defaultdict(deque)
-
-        self._enable_activation_logging()
+        self._param_updates: Optional[Dict[str, torch.Tensor]] = None
 
     def log_scalars(
         self,
@@ -136,6 +119,89 @@ class LogWriter(Registrable):
         Defines how batch inputs are logged. This is called once at the start of each epoch.
         """
         pass
+
+    def close(self) -> None:
+        """
+        Called at the end of training to remove any module hooks and close out any
+        other logging resources.
+        """
+        for handle in self._module_hook_handles:
+            handle.remove()
+
+    def on_start(
+        self, trainer: "GradientDescentTrainer", is_primary: bool = True, **kwargs
+    ) -> None:
+        self.trainer = trainer
+        if is_primary:
+            self._enable_activation_logging()
+
+    def on_batch(
+        self,
+        trainer: "GradientDescentTrainer",
+        batch_inputs: List[TensorDict],
+        batch_outputs: List[Dict[str, Any]],
+        batch_metrics: Dict[str, Any],
+        epoch: int,
+        batch_number: int,
+        is_training: bool,
+        is_primary: bool = True,
+        batch_grad_norm: Optional[float] = None,
+        **kwargs,
+    ) -> None:
+        if not is_training and not is_primary:
+            return None
+        assert self.trainer is not None
+
+        if self._should_log_distributions_this_batch():
+            assert self._param_updates is not None
+            for name, param in trainer.model.named_parameters():
+                self._param_updates[name].sub_(param.detach().cpu())
+        else:
+            self._param_updates = None
+
+        self.log_batch(
+            batch_grad_norm,
+            batch_metrics,
+            batch_inputs,
+            self._param_updates,
+            batch_number,
+        )
+
+        if self._should_log_distributions_next_batch():
+            self._param_updates = {
+                name: param.detach().cpu().clone()
+                for name, param in trainer.model.named_parameters()
+            }
+
+    def on_epoch(
+        self,
+        trainer: "GradientDescentTrainer",
+        metrics: Dict[str, Any],
+        epoch: int,
+        is_primary: bool = True,
+        **kwargs,
+    ) -> None:
+        if not is_primary:
+            return None
+        assert self.trainer is not None
+
+        train_metrics, val_metrics = get_train_and_validation_metrics(metrics)
+        self.log_epoch(
+            train_metrics,
+            val_metrics,
+            epoch,
+        )
+
+    def on_end(
+        self,
+        trainer: "GradientDescentTrainer",
+        metrics: Dict[str, Any] = None,
+        epoch: int = None,
+        is_primary: bool = True,
+        **kwargs,
+    ) -> None:
+        if is_primary:
+            self.close()
 
     def log_batch(
         self,
@@ -180,7 +246,7 @@ class LogWriter(Registrable):
                 log_prefix="train",
             )
 
-        if self.should_log_distributions_this_batch():
+        if self._should_log_distributions_this_batch():
             assert param_updates is not None
             self._log_distributions()
             self._log_gradient_updates(param_updates)
@@ -189,15 +255,13 @@ class LogWriter(Registrable):
             # We're assuming here that `log_batch` will get called every batch, and only every
             # batch.  This is true with our current usage of this code (version 1.0); if that
             # assumption becomes wrong, this code will break.
-            batch_group_size = sum(training_util.get_batch_size(batch) for batch in batch_group)  # type: ignore
+            batch_group_size = sum(get_batch_size(batch) for batch in batch_group)  # type: ignore
             self._cumulative_batch_group_size += batch_group_size
             if batch_number % self._batch_size_interval == 0:
                 average = self._cumulative_batch_group_size / batch_number
                 self.log_scalars(
                     {"batch_size": batch_group_size, "mean_batch_size": average}, log_prefix="train"
                 )
-
-        self._batch_num_total += 1
 
     def log_epoch(
         self,
@@ -224,16 +288,16 @@ class LogWriter(Registrable):
             epoch=epoch,
         )
 
-    def should_log_distributions_next_batch(self) -> bool:
+    def _should_log_distributions_next_batch(self) -> bool:
         return (
             self._distribution_interval is not None
-            and (self._batch_num_total + 2) % self._distribution_interval == 0
+            and (self.trainer._batch_num_total + 1) % self._distribution_interval == 0  # type: ignore[union-attr]
         )
 
-    def should_log_distributions_this_batch(self) -> bool:
+    def _should_log_distributions_this_batch(self) -> bool:
         return (
             self._distribution_interval is not None
-            and (self._batch_num_total + 1) % self._distribution_interval == 0
+            and self.trainer._batch_num_total % self._distribution_interval == 0  # type: ignore[union-attr]
         )
 
     def _enable_activation_logging(self) -> None:
@@ -242,19 +306,19 @@ class LogWriter(Registrable):
             # a hook on forward to capture the output tensors.
             # This uses a closure to determine whether to log the activations,
             # since we don't want them on every call.
-            for _, module in self._model.named_modules():
+            for _, module in self.trainer.model.named_modules():  # type: ignore[union-attr]
                 if not getattr(module, "should_log_activations", False):
                     # skip it
                     continue
 
                 def hook(module_, inputs, outputs):
-                    if self.should_log_distributions_this_batch():
+                    if self._should_log_distributions_this_batch():
                         self._log_activation_distribution(outputs, str(module_.__class__))
 
                 self._module_hook_handles.append(module.register_forward_hook(hook))
 
     def _should_log_this_batch(self) -> bool:
-        return self._batch_num_total % self._summary_interval == 0
+        return self.trainer._batch_num_total % self._summary_interval == 0  # type: ignore[union-attr]
 
     def _log_activation_distribution(self, outputs: Any, module_name: str) -> None:
         activations_to_log: Dict[str, torch.Tensor] = {}
@@ -281,7 +345,7 @@ class LogWriter(Registrable):
         gradient_mean_scalars: Dict[str, float] = {}
         gradient_std_scalars: Dict[str, float] = {}
         # Log parameter values to TensorBoard
-        for name, param in self._model.named_parameters():
+        for name, param in self.trainer.model.named_parameters():  # type: ignore[union-attr]
             if param.data.numel() > 0:
                 parameter_mean_scalars[name] = param.data.mean().item()
             if param.data.numel() > 1:
@@ -312,8 +376,8 @@ class LogWriter(Registrable):
         # optimizer stores lr info keyed by parameter tensor
         # we want to log with parameter name
         lr_scalars: Dict[str, float] = {}
-        names = {param: name for name, param in self._model.named_parameters()}
-        for group in self._optimizer.param_groups:
+        names = {param: name for name, param in self.trainer.model.named_parameters()}  # type: ignore[union-attr]
+        for group in self.trainer.optimizer.param_groups:  # type: ignore[union-attr]
             if "lr" not in group:
                 continue
             rate = group["lr"]
@@ -330,27 +394,21 @@ class LogWriter(Registrable):
         if not self._distribution_parameters:
             # Avoiding calling this every batch.  If we ever use two separate models with a single
             # writer, this is wrong, but I doubt that will ever happen.
-            self._distribution_parameters = set(self._model.get_parameters_for_histogram_logging())
+            self._distribution_parameters = set(
+                self.trainer.model.get_parameters_for_histogram_logging()  # type: ignore[union-attr]
+            )
         parameters_to_log: Dict[str, torch.Tensor] = {}
-        for name, param in self._model.named_parameters():
+        for name, param in self.trainer.model.named_parameters():  # type: ignore[union-attr]
             if name in self._distribution_parameters:
                 parameters_to_log[name] = param
         self.log_tensors(parameters_to_log, log_prefix="parameter_histogram")
 
     def _log_gradient_updates(self, param_updates: Dict[str, torch.Tensor]) -> None:
         gradient_update_scalars: Dict[str, float] = {}
-        for name, param in self._model.named_parameters():
+        for name, param in self.trainer.model.named_parameters():  # type: ignore[union-attr]
             update_norm = torch.norm(param_updates[name].view(-1))
             param_norm = torch.norm(param.view(-1)).cpu()
             gradient_update_scalars[name] = (
-                update_norm / (param_norm + nn_util.tiny_value_of_dtype(param_norm.dtype))
+                update_norm / (param_norm + tiny_value_of_dtype(param_norm.dtype))
             ).item()
         self.log_scalars(gradient_update_scalars, log_prefix="gradient_update")
-
-    def close(self) -> None:
-        """
-        Called at the end of training to remove any module hooks and close out any
-        other logging resources.
-        """
-        for handle in self._module_hook_handles:
-            handle.remove()
