+++ allennlp/training/callbacks/tensorboard.py
@@ -1,49 +1,31 @@
-import warnings
-from typing import Dict, Optional, Union
-import logging
 import os
+from typing import Dict, Union, Optional
 
 from overrides import overrides
 from tensorboardX import SummaryWriter
 import torch
 
-from allennlp.models.model import Model
-from allennlp.training.log_writer import LogWriter
-from allennlp.training.optimizers import Optimizer
+from allennlp.training.callbacks.callback import TrainerCallback
+from allennlp.training.callbacks.log_writer import LogWriterCallback
 
 
-logger = logging.getLogger(__name__)
-
-
-@LogWriter.register("tensorboard")
-class TensorBoardWriter(LogWriter):
+@TrainerCallback.register("tensorboard")
+class TensorBoardCallback(LogWriterCallback):
     """
-    A `LogWriter` that writes to TensorBoard.
+    A callback that writes training statistics/metrics to TensorBoard.
     """
 
     def __init__(
         self,
         serialization_dir: str,
-        model: Model,
-        optimizer: Optimizer,
         summary_interval: int = 100,
         distribution_interval: Optional[int] = None,
         batch_size_interval: Optional[int] = None,
         should_log_parameter_statistics: bool = True,
         should_log_learning_rate: bool = False,
-        **kwargs,
     ) -> None:
-        if "histogram_interval" in kwargs:
-            warnings.warn(
-                "`histogram_interval` is deprecated." "Use `distribution_interval` instead.",
-                DeprecationWarning,
-            )
-            distribution_interval = kwargs.pop("histogram_interval")
-
         super().__init__(
             serialization_dir,
-            model,
-            optimizer,
             summary_interval=summary_interval,
             distribution_interval=distribution_interval,
             batch_size_interval=batch_size_interval,
@@ -53,10 +35,10 @@ class TensorBoardWriter(LogWriter):
 
         # Create log directories prior to creating SummaryWriter objects
         # in order to avoid race conditions during distributed training.
-        train_ser_dir = os.path.join(self._serialization_dir, "log", "train")
+        train_ser_dir = os.path.join(self.serialization_dir, "log", "train")
         os.makedirs(train_ser_dir, exist_ok=True)
         self._train_log = SummaryWriter(train_ser_dir)
-        val_ser_dir = os.path.join(self._serialization_dir, "log", "validation")
+        val_ser_dir = os.path.join(self.serialization_dir, "log", "validation")
         os.makedirs(val_ser_dir, exist_ok=True)
         self._validation_log = SummaryWriter(val_ser_dir)
 
@@ -67,7 +49,7 @@ class TensorBoardWriter(LogWriter):
         log_prefix: str = "",
         epoch: Optional[int] = None,
     ) -> None:
-        timestep = epoch if epoch is not None else self._batch_num_total
+        timestep = epoch if epoch is not None else self.trainer._batch_num_total  # type: ignore[union-attr]
         log = self._train_log if not log_prefix.startswith("validation") else self._validation_log
         for key, value in scalars.items():
             name = f"{log_prefix}/{key}" if log_prefix else key
@@ -77,7 +59,7 @@ class TensorBoardWriter(LogWriter):
     def log_tensors(
         self, tensors: Dict[str, torch.Tensor], log_prefix: str = "", epoch: Optional[int] = None
     ) -> None:
-        timestep = epoch if epoch is not None else self._batch_num_total
+        timestep = epoch if epoch is not None else self.trainer._batch_num_total  # type: ignore[union-attr]
         log = self._train_log if not log_prefix.startswith("validation") else self._validation_log
         for key, values in tensors.items():
             name = f"{log_prefix}/{key}" if log_prefix else key
