+++ allennlp/training/trainer.py
@@ -23,35 +23,15 @@ from allennlp.common.checks import ConfigurationError, check_for_gpu
 from allennlp.data import DataLoader, TensorDict
 from allennlp.models.model import Model
 from allennlp.training import util as training_util
+from allennlp.training.callbacks import TrainerCallback, SanityChecksCallback
 from allennlp.training.checkpointer import Checkpointer
 from allennlp.training.learning_rate_schedulers import LearningRateScheduler
 from allennlp.training.metric_tracker import MetricTracker
 from allennlp.training.momentum_schedulers import MomentumScheduler
 from allennlp.training.moving_average import MovingAverage
 from allennlp.training.optimizers import Optimizer
-from allennlp.training.tensorboard_writer import TensorBoardWriter
-from allennlp.training.wandb_writer import WandBWriter
-from allennlp.sanity_checks.normalization_bias_verification import NormalizationBiasVerification
-from allennlp.training.log_writer import LogWriter
 
 logger = logging.getLogger(__name__)
-
-
-def _get_train_and_validation_metrics(metrics: Dict) -> Tuple[Dict[str, Any], Dict[str, Any]]:
-    """
-    Utility function to separate out train_metrics and val_metrics.
-    """
-    train_metrics: Dict[str, Any] = {}
-    val_metrics: Dict[str, Any] = {}
-    for key, value in metrics.items():
-        if key.startswith("training_"):
-            key = key.replace("training_", "", 1)
-            if key not in {"duration", "start_epoch", "epochs"}:
-                train_metrics[key] = value
-        elif key.startswith("validation_"):
-            key = key.replace("validation_", "", 1)
-            val_metrics[key] = value
-    return train_metrics, val_metrics
 
 
 class Trainer(Registrable):
@@ -127,360 +107,6 @@ class Trainer(Registrable):
         for checkpointing.  This is used, e.g., for moving averages of model weights.
         """
         raise NotImplementedError
-
-
-class TrainerCallback(Registrable):
-    """
-    A general callback object that handles multiple events.
-
-    This class has `on_batch`, `on_epoch`, and `on_end` methods, corresponding to
-    each callback type. Each one receives the state of the wrapper object as `self`.
-    This enables easier state sharing between related callbacks.
-
-    Also, this callback type is instantiated with `serialization_dir` and `on_start` is called
-    with the trainer instance as an argument. This might be handy in case of callback logging
-    and saving its own files next to the config/checkpoints/logs/etc.
-    """
-
-    def __init__(self, serialization_dir: str) -> None:
-        self.serialization_dir = serialization_dir
-        self.trainer: Optional["GradientDescentTrainer"] = None
-
-    def on_start(
-        self, trainer: "GradientDescentTrainer", is_primary: bool = True, **kwargs
-    ) -> None:
-        """
-        This callback hook is called before the training is started.
-        """
-        self.trainer = trainer
-
-    def on_batch(
-        self,
-        trainer: "GradientDescentTrainer",
-        batch_inputs: List[TensorDict],
-        batch_outputs: List[Dict[str, Any]],
-        batch_metrics: Dict[str, Any],
-        epoch: int,
-        batch_number: int,
-        is_training: bool,
-        is_primary: bool = True,
-        batch_grad_norm: Optional[float] = None,
-        **kwargs,
-    ) -> None:
-        """
-        This callback hook is called after the end of each batch.
-        """
-        pass
-
-    def on_epoch(
-        self,
-        trainer: "GradientDescentTrainer",
-        metrics: Dict[str, Any],
-        epoch: int,
-        is_primary: bool = True,
-        **kwargs,
-    ) -> None:
-        """
-        This callback hook is called after the end of each epoch.
-        """
-        pass
-
-    def on_end(
-        self,
-        trainer: "GradientDescentTrainer",
-        metrics: Dict[str, Any] = None,
-        epoch: int = None,
-        is_primary: bool = True,
-        **kwargs,
-    ) -> None:
-        """
-        This callback hook is called after the final training epoch.
-        """
-        pass
-
-
-TrainerCallback.register("null")(TrainerCallback)
-
-
-@TrainerCallback.register("sanity_checks")
-class SanityCheckCallback(TrainerCallback):
-    """
-    Performs model sanity checks.
-
-    Checks performed:
-
-    * `NormalizationBiasVerification` for detecting invalid combinations of
-       bias and normalization layers.
-       See `allennlp.sanity_checks.normalization_bias_verification` for more details.
-
-    Note: Any new sanity checks should also be added to this callback.
-    """
-
-    def on_start(
-        self, trainer: "GradientDescentTrainer", is_primary: bool = True, **kwargs
-    ) -> None:
-        self.trainer = trainer
-        if is_primary:
-            self._verification = NormalizationBiasVerification(self.trainer._pytorch_model)
-            # Register the hooks that perform the verification before training starts.
-            self._verification.register_hooks()
-
-    def on_batch(
-        self,
-        trainer: "GradientDescentTrainer",
-        batch_inputs: List[TensorDict],
-        batch_outputs: List[Dict[str, Any]],
-        batch_metrics: Dict[str, Any],
-        epoch: int,
-        batch_number: int,
-        is_training: bool,
-        is_primary: bool = True,
-        batch_grad_norm: Optional[float] = None,
-        **kwargs,
-    ) -> None:
-        if not is_primary:
-            return None
-
-        # We destroy the hooks after the first batch, since we only want to
-        # perform this check once.
-        if epoch == 0 and batch_number == 1 and is_training:
-            self._verification.destroy_hooks()
-            detected_pairs = self._verification.collect_detections()
-            assert (
-                len(detected_pairs) == 0
-            ), "The NormalizationBiasVerification check failed. See logs for more details."
-
-
-@TrainerCallback.register("log_writer")
-class LogWriterCallback(TrainerCallback):
-    """
-    Log training statistics and metrics using a `LogWriter`.
-    """
-
-    def __init__(
-        self,
-        serialization_dir: str,
-        log_writer: Lazy[LogWriter],
-    ) -> None:
-        super().__init__(serialization_dir)
-        self._log_writer_constructor = log_writer
-        self._log_writer: Optional[LogWriter] = None
-        self._param_updates: Optional[Dict[str, torch.Tensor]] = None
-
-    def on_start(
-        self, trainer: "GradientDescentTrainer", is_primary: bool = True, **kwargs
-    ) -> None:
-        self.trainer = trainer
-        if is_primary:
-            self._log_writer = self._log_writer_constructor.construct(
-                serialization_dir=self.serialization_dir,
-                model=self.trainer.model,
-                optimizer=self.trainer.optimizer,
-            )
-
-    def on_batch(
-        self,
-        trainer: "GradientDescentTrainer",
-        batch_inputs: List[TensorDict],
-        batch_outputs: List[Dict[str, Any]],
-        batch_metrics: Dict[str, Any],
-        epoch: int,
-        batch_number: int,
-        is_training: bool,
-        is_primary: bool = True,
-        batch_grad_norm: Optional[float] = None,
-        **kwargs,
-    ) -> None:
-        if not is_training and not is_primary:
-            return None
-        assert self._log_writer is not None
-
-        if self._log_writer.should_log_distributions_this_batch():
-            assert self._param_updates is not None
-            for name, param in trainer.model.named_parameters():
-                self._param_updates[name].sub_(param.detach().cpu())
-        else:
-            self._param_updates = None
-        # Need to call this before .log_batch(), since .log_batch() updates its
-        # internal batch num counter.
-        should_log_distributions_next_batch = self._log_writer.should_log_distributions_next_batch()
-
-        self._log_writer.log_batch(
-            batch_grad_norm,
-            batch_metrics,
-            batch_inputs,
-            self._param_updates,
-            batch_number,
-        )
-
-        if should_log_distributions_next_batch:
-            self._param_updates = {
-                name: param.detach().cpu().clone()
-                for name, param in trainer.model.named_parameters()
-            }
-
-    def on_epoch(
-        self,
-        trainer: "GradientDescentTrainer",
-        metrics: Dict[str, Any],
-        epoch: int,
-        is_primary: bool = True,
-        **kwargs,
-    ) -> None:
-        if not is_primary:
-            return None
-        assert self._log_writer is not None
-
-        train_metrics, val_metrics = _get_train_and_validation_metrics(metrics)
-        self._log_writer.log_epoch(
-            train_metrics,
-            val_metrics,
-            epoch,
-        )
-
-    def on_end(
-        self,
-        trainer: "GradientDescentTrainer",
-        metrics: Dict[str, Any] = None,
-        epoch: int = None,
-        is_primary: bool = True,
-        **kwargs,
-    ) -> None:
-        if is_primary:
-            assert self._log_writer is not None
-            self._log_writer.close()
-
-    @classmethod
-    def tensorboard(
-        cls,
-        serialization_dir: str,
-        tensorboard_writer: Lazy[TensorBoardWriter] = Lazy(TensorBoardWriter),
-    ) -> "LogWriterCallback":
-        return cls(serialization_dir, tensorboard_writer)  # type: ignore[arg-type]
-
-    @classmethod
-    def wandb(
-        cls, serialization_dir: str, wandb_writer: Lazy[WandBWriter] = Lazy(WandBWriter)
-    ) -> "LogWriterCallback":
-        return cls(serialization_dir, wandb_writer)  # type: ignore[arg-type]
-
-
-TrainerCallback.register("tensorboard", constructor="tensorboard")(LogWriterCallback)
-TrainerCallback.register("wandb", constructor="wandb")(LogWriterCallback)
-
-
-@TrainerCallback.register("console_logger")
-class ConsoleLoggerCallback(TrainerCallback):
-    def __init__(
-        self,
-        serialization_dir: str,
-        should_log_inputs: bool = False,
-    ) -> None:
-        super().__init__(serialization_dir)
-        self._should_log_inputs = should_log_inputs
-
-    def on_batch(
-        self,
-        trainer: "GradientDescentTrainer",
-        batch_inputs: List[TensorDict],
-        batch_outputs: List[Dict[str, Any]],
-        batch_metrics: Dict[str, Any],
-        epoch: int,
-        batch_number: int,
-        is_training: bool,
-        is_primary: bool = True,
-        batch_grad_norm: Optional[float] = None,
-        **kwargs,
-    ) -> None:
-
-        if not is_primary:
-            return None
-
-        # We only want to do this for the first batch in the first epoch.
-        if batch_number == 1 and epoch == 0 and self._should_log_inputs:
-            logger.info("Batch inputs")
-            for b, batch in enumerate(batch_inputs):
-                self._log_fields(batch, log_prefix="batch_input")  # type: ignore
-
-    def _log_fields(self, fields: Dict, log_prefix: str = ""):
-        for key, val in fields.items():
-            key = log_prefix + "/" + key
-            if isinstance(val, dict):
-                self._log_fields(val, key)
-            elif isinstance(val, torch.Tensor):
-                torch.set_printoptions(threshold=2)
-                logger.info("%s (Shape: %s)\n%s", key, " x ".join([str(x) for x in val.shape]), val)
-                torch.set_printoptions(threshold=1000)
-            elif isinstance(val, List):
-                logger.info('Field : "%s" : (Length %d of type "%s")', key, len(val), type(val[0]))
-            elif isinstance(val, str):
-                logger.info('Field : "{}" : "{:20.20} ..."'.format(key, val))
-            else:
-                logger.info('Field : "%s" : %s', key, val)
-
-    def on_epoch(
-        self,
-        trainer: "GradientDescentTrainer",
-        metrics: Dict[str, Any],
-        epoch: int,
-        is_primary: bool = True,
-        **kwargs,
-    ) -> None:
-        if not is_primary:
-            return None
-
-        train_metrics, val_metrics = _get_train_and_validation_metrics(metrics)
-
-        metric_names = set(train_metrics.keys())
-        if val_metrics is not None:
-            metric_names.update(val_metrics.keys())
-        val_metrics = val_metrics or {}
-
-        dual_message_template = "%s |  %8.3f  |  %8.3f"
-        no_val_message_template = "%s |  %8.3f  |  %8s"
-        no_train_message_template = "%s |  %8s  |  %8.3f"
-        header_template = "%s |  %-10s"
-        name_length = max(len(x) for x in metric_names)
-        logger.info(header_template, "Training".rjust(name_length + 13), "Validation")
-
-        for name in sorted(metric_names):
-            train_metric = train_metrics.get(name)
-            val_metric = val_metrics.get(name)
-
-            if val_metric is not None and train_metric is not None:
-                logger.info(
-                    dual_message_template, name.ljust(name_length), train_metric, val_metric
-                )
-            elif val_metric is not None:
-                logger.info(no_train_message_template, name.ljust(name_length), "N/A", val_metric)
-            elif train_metric is not None:
-                logger.info(no_val_message_template, name.ljust(name_length), train_metric, "N/A")
-
-
-@TrainerCallback.register("track_epoch_callback")
-class TrackEpochCallback(TrainerCallback):
-    """
-    A callback that you can pass to the `GradientDescentTrainer` to access the current epoch number
-    in your model during training. This callback sets `model.epoch`, which can be read inside of
-    `model.forward()`. We set `model.epoch = epoch + 1` which now denotes the number of
-    completed epochs at a given training state.
-    """
-
-    def on_start(
-        self, trainer: "GradientDescentTrainer", is_primary: bool = True, **kwargs
-    ) -> None:
-        super().on_start(trainer, is_primary)
-        trainer.model.epoch = 0
-
-    def on_epoch(
-        self,
-        trainer: "GradientDescentTrainer",
-        metrics: Dict[str, Any],
-        epoch: int,
-        is_primary: bool = True,
-        **kwargs,
-    ) -> None:
-        trainer.model.epoch = epoch + 1
 
 
 @Trainer.register("gradient_descent", constructor="from_partial_objects")
@@ -1421,7 +1047,7 @@ class GradientDescentTrainer(Trainer):
         # destroyed after the first run, therefore, the second instance of the
         # callback will do nothing.
         if run_sanity_check:
-            callbacks_.append(SanityCheckCallback(serialization_dir=serialization_dir))
+            callbacks_.append(SanityChecksCallback(serialization_dir=serialization_dir))
 
         return cls(
             model,
